requirements.txt
python-telegram-bot[job-queue]>=21.0
httpx>=0.28.0
pydantic-settings>=2.13.0
python-dotenv>=1.0.0
pillow>=11.0
aiohttp>=3.9.0
